requirements.txt
notebook
ipykernel
matplotlib
networkx
qiskit
numpy
pandas
openpyxl
PyYAML